+++ CHANGELOG.md
@@ -15,3 +15,19 @@
 - Заменил во всех классах enum на enum class
 - Добавил для всех моих классов namespace 'WGP'
 - В старых классах card, deck и klondikeField изменил стиль названия с snake_case на camelCase
+
+## [Version: alpha 0.0.4.1] - 10.10.2025
+
+### Изменения
+
+- добавил функции setCardsRandomRotation для класса deckRenderer.
+- добавил функцию flip для класса deck.
+- переделал функцию openCardFromStock так, чтобы в _stockPile хранились только FaceDown карты, а в _wastePile только FaceUp.
+- переделал логику update у класса cardRenderer так, чтобы при каждом  draw не вызывался update, а только когда необходимо.
+- переделал функцию update у класса deckRenderer, заменив вычисления на функцию getCoordinatesVectorFromAngle.
+- переделал конструкторы класса deckRenderer так, чтобы инициализация полей проходила не в сторонних функциях, а в теле конструктора.
+- переделал функции рандома класса card с использованием std::mt19937.
+- Добавил классу deckRenderer стиль колоды (FromPosition, FromAngleSpacing) для корректного изменения характеристик расположения колоды.
+- Добавил метод initialize классу cardRenderer, который инициализирует поля класса из одной функции.
+- Добавил методы initializeFromAngle и initializeFromPosition классу deckRenderer, которые инициализируют поля класса из одной функции.
+- Добавил метод spacingFromLength класса deckRenderer.